#110 restructure simulation app so hopefully will work on heroku
