+++ pkpdapp/pkpdapp/erlotinib/_mechanistic_models.py
@@ -467,9 +467,6 @@ class PharmacokineticModel(MechanisticModel):
                              str(compartment) + '>.')
         comp = model.get(compartment, class_filter=myokit.Component)
 
-        print('model has variables')
-        for v in comp.variables():
-            print(v.name())
         if not comp.has_variable(amount_var):
             raise ValueError('The drug amount variable <' + str(amount_var) +
                              '> could not '
