+++ pkpdapp/pkpdapp/templates/dataset_detail.html
@@ -20,7 +20,7 @@ copyright notice and full license details.
 {% endfor %}
 </ul>
 
-{% plotly_app_bootstrap name="dataset_view" aspect="16by9"%}
+{% plotly_app_bootstrap name="data_view_app" aspect="16by9"%}
 
 <h3>Protocols</h2>
 <table style="width:100%">
