+++ pkpdapp/pkpdapp/templates/dosed_pharmacokinetic_detail.html
@@ -27,6 +27,6 @@ copyright notice and full license details.
 <ul>
 </div>
 <div class="container-xl">
-  {% plotly_app_bootstrap name="dosed_parmokinetic_view" %}
+  {% plotly_app_bootstrap name="model_view" %}
 </div>
 {% endblock %}
