+++ pkpdapp/pkpdapp/views/pkpd_model.py
@@ -16,122 +16,32 @@ from pkpdapp.models import (PharmacokineticModel, DosedPharmacokineticModel,
                             PharmacodynamicModel, Biomarker, BiomarkerType,
                             Dose)
 import pandas as pd
-from pkpdapp.dash_apps.simulation import PKSimulationApp, PDSimulationApp
+from pkpdapp.dash_apps.model_view import ModelViewState
 import pkpdapp.erlotinib as erlo
 from dash.dependencies import Input, Output
 import dash
 from myokit.formats.sbml import SBMLParsingError
 
 
-def create_dash_app(model, project):
-    # create dash app
-
-    if isinstance(model, PharmacodynamicModel):
-        app = PDSimulationApp(name='parmacodynamic_view')
-        sbml_str = model.sbml.encode('utf-8')
-        try:
-            erlo_m = erlo.PharmacodynamicModel(sbml_str)
-            param_names = erlo_m.parameters()
-            param_names_dict = {}
-            for p in param_names:
-                param_names_dict[p] = p.replace('.', '_')
-            erlo_m.set_parameter_names(names=param_names_dict)
-            app.add_model(erlo_m, model.name, use=True)
-        except SBMLParsingError:
-            pass
-
+def create_model_view_state(model, project):
+    # create dash state
+    state = ModelViewState()
+    is_pk = isinstance(model, DosedPharmacokineticModel)
+    if is_pk:
+        state.add_model(
+            model.pharmacokinetic_model.sbml,
+            model.pharmacokinetic_model.name, is_pk=is_pk, use=True
+        )
+        state.set_administration(model.dose_compartment)
+        events = [
+            (d.amount, d.start_time, d.duration)
+            for d in Dose.objects.filter(protocol=model.protocol)
+        ]
+        state.set_dosing_events(events)
     else:
-        app = PKSimulationApp(name='dosed_parmokinetic_view')
-        # get model sbml strings and add erlo models
-        m = model.pharmacokinetic_model
-        sbml_str = m.sbml.encode('utf-8')
-        try:
-            erlo_m = erlo.PharmacokineticModel(sbml_str)
-            param_names = erlo_m.parameters()
-            param_names_dict = {}
-            for p in param_names:
-                param_names_dict[p] = p.replace('.', '_')
-            erlo_m.set_parameter_names(names=param_names_dict)
-            app.add_model(erlo_m, m.name, use=True)
-            app.set_administration(model.dose_compartment)
-            events = [
-                (d.amount, d.start_time, d.duration)
-                for d in Dose.objects.filter(protocol=model.protocol)
-            ]
-            app.set_dosing_events(events)
-        except SBMLParsingError:
-            pass
+        state.add_model(model.sbml, model.name, is_pk=is_pk, use=True)
 
-    # add datasets
-    if project:
-        for i, d in enumerate(project.datasets.all()):
-            biomarker_types = BiomarkerType.objects.filter(dataset=d)
-            biomarkers = Biomarker.objects\
-                .select_related('biomarker_type__name')\
-                .filter(biomarker_type__in=biomarker_types)
-
-            # convert to pandas dataframe with the column names expected
-            df = pd.DataFrame(
-                list(
-                    biomarkers.values('time', 'subject_id',
-                                      'biomarker_type__name', 'value')))
-            df.rename(columns={
-                'subject_id': 'ID',
-                'time': 'Time',
-                'biomarker_type__name': 'Biomarker',
-                'value': 'Measurement'
-            }, inplace=True)
-
-            app.add_data(df, d.name, use=False)
-
-    # generate dash app
-    app.set_layout()
-
-    # we need slider ids for callback, count the number of parameters for
-    # each model so we know what parameter in the list corresponds to which
-    # model
-    sliders = app.slider_ids()
-    n_params = [len(s) for s in sliders]
-    offsets = [0]
-    for i in range(1, len(n_params)):
-        offsets.append(offsets[i - 1] + n_params[i])
-
-    # Define simulation callbacks
-    @app.app.callback(
-        Output('fig', 'figure'), [Input(s, 'value') for s in sum(sliders, [])],
-        [Input('dataset-select', 'value'),
-         Input('biomarker-select', 'value')])
-    def update_simulation(*args):
-        """
-        if the models, datasets or biomarkers are
-        changed then regenerate the figure entirely
-
-        if a slider is moved, determine the relevent model based on the id
-        name, then update that particular simulation
-        """
-        ctx = dash.callback_context
-        cid = None
-        if ctx.triggered:
-            cid = ctx.triggered[0]['prop_id'].split('.')[0]
-        print('ARGS', args)
-
-        if cid == 'dataset-select':
-            print('update datasets')
-            app.set_used_datasets(args[-2])
-            return app.create_figure()
-        elif cid == 'biomarker-select':
-            app.set_used_biomarker(args[-1])
-            return app.create_figure()
-        elif cid is not None:
-            model_index = int(cid[:2])
-            n = n_params[model_index]
-            offset = offsets[model_index]
-            parameters = args[offset:offset + n]
-            return app.update_simulation(model_index, parameters)
-
-        return app._fig._fig
-
-    return app
+    return state
 
 
 class PharmacodynamicModelDetailView(LoginRequiredMixin, DetailView):
@@ -139,8 +49,13 @@ class PharmacodynamicModelDetailView(LoginRequiredMixin, DetailView):
     template_name = 'pd_model_detail.html'
 
     def get(self, request, *args, **kwargs):
-        self._app = create_dash_app(self.get_object(),
-                                    self.request.user.profile.selected_project)
+        session = request.session
+        session['django_plotly_dash'] = {
+            'model_view': create_model_view_state(
+                self.get_object(),
+                self.request.user.profile.selected_project
+            ).to_json()
+        }
         return super().get(request)
 
 
@@ -184,8 +99,13 @@ class DosedPharmacokineticModelDetail(LoginRequiredMixin, DetailView):
     model = DosedPharmacokineticModel
 
     def get(self, request, *args, **kwargs):
-        self._app = create_dash_app(self.get_object(),
-                                    self.request.user.profile.selected_project)
+        session = request.session
+        session['django_plotly_dash'] = {
+            'model_view': create_model_view_state(
+                self.get_object(),
+                self.request.user.profile.selected_project
+            ).to_json()
+        }
         return super().get(request)
